+++ src/napari/_vispy/_tests/test_vispy_labels_layer.py
@@ -3,10 +3,32 @@ import pytest
 import zarr
 from qtpy.QtCore import QCoreApplication
 
+import napari._vispy.layers.base as vispy_base
+import napari._vispy.layers.labels as vispy_labels
 from napari._tests.utils import skip_local_popups, skip_on_win_ci
+from napari._vispy.layers.labels import (
+    DirectLabelVispyColormap,
+    LabelVispyColormap,
+    VispyLabelsLayer,
+)
+from napari._vispy.utils.qt_font import FontInfo
 from napari._vispy.visuals.volume import Volume as VolumeNode
+from napari.layers import Labels
 from napari.layers.labels._labels_constants import IsoCategoricalGradientMode
+from napari.settings import get_settings
+from napari.utils.colormaps import DirectLabelColormap
 from napari.utils.interactions import mouse_press_callbacks
+
+
+@pytest.fixture
+def _mock_max_texture_sizes(monkeypatch):
+    max_texture_sizes = (16384, 2048)
+    monkeypatch.setattr(
+        vispy_base, 'get_max_texture_sizes', lambda: max_texture_sizes
+    )
+    monkeypatch.setattr(
+        vispy_labels, 'get_max_texture_sizes', lambda: max_texture_sizes
+    )
 
 
 def make_labels_layer(array_type, shape):
@@ -30,6 +52,60 @@ def make_labels_layer(array_type, shape):
         pytest.fail("array_type must be 'numpy', 'zarr', or 'tensorstore'")
 
     return labels
+
+
+@pytest.mark.usefixtures('_fresh_settings', '_mock_max_texture_sizes')
+def test_colormap_rebuilt_when_slice_dtypes_change():
+    get_settings().experimental.async_ = True
+    data = np.zeros((8, 8), dtype=np.uint32)
+    data[0, 0] = 100_000
+    layer = Labels(data)
+    visual = VispyLabelsLayer(layer, font_info=FontInfo())
+
+    try:
+        layer.colormap = DirectLabelColormap(
+            color_dict={
+                None: [0, 0, 0, 0],
+                0: [0, 0, 0, 0],
+                1: [0, 0.25, 1, 1],
+                100_000: [1, 0, 0, 1],
+            }
+        )
+        assert layer._slice.empty
+        # The empty placeholder uses uint8 for both raw and view, so the
+        # colormap is initially configured for that pair. Retaining this state
+        # for the real (uint32, uint8) slice would produce stale label colors.
+        # _on_data_change detects this transition and rebuilds the colormap.
+        assert visual._colormap_dtypes == (np.dtype('uint8'),) * 2
+        assert isinstance(visual.node.cmap, LabelVispyColormap)
+
+        layer.set_view_slice()
+        layer.events.set_data()
+
+        # set_data must rebuild the colormap instead of leaving the
+        # LabelVispyColormap configured for the placeholder.
+        assert visual._colormap_dtypes == (
+            np.dtype('uint32'),
+            np.dtype('uint8'),
+        )
+        assert isinstance(visual.node.cmap, DirectLabelVispyColormap)
+    finally:
+        visual.close()
+
+
+@pytest.mark.usefixtures('_mock_max_texture_sizes')
+def test_colormap_not_rebuilt_when_slice_dtypes_are_unchanged():
+    layer = Labels(np.zeros((8, 8), dtype=np.uint32))
+    visual = VispyLabelsLayer(layer, font_info=FontInfo())
+
+    try:
+        colormap = visual.node.cmap
+
+        layer.events.set_data()
+
+        assert visual.node.cmap is colormap
+    finally:
+        visual.close()
 
 
 @skip_local_popups
